+++ packages/overlayscrollbars/src/setups/scrollbarsSetup/scrollbarsSetup.events.ts
@@ -44,7 +44,11 @@ export type ScrollbarsSetupEvents = (
 export const createScrollbarsSetupEvents = (
   options: ReadonlyOptions,
   structureSetupElements: StructureSetupElementsObj,
-  structureSetupState: StructureSetupState
+  structureSetupState: StructureSetupState,
+  scrollbarHandlePointerInteraction: (
+    event: PointerEvent,
+    scrollbarsAddRemoveClass: ScrollbarsSetupElementsObj['_scrollbarsAddRemoveClass']
+  ) => void
 ): ScrollbarsSetupEvents => {
   const { _host, _scrollOffsetElement, _documentElm } = structureSetupElements;
 
@@ -211,6 +215,9 @@ export const createScrollbarsSetupEvents = (
         },
         { _passive: false, _capture: true }
       ),
+      addEventListener(_handle, 'pointermove pointerleave', (event: PointerEvent) => {
+        scrollbarHandlePointerInteraction(event, scrollbarsAddRemoveClass);
+      }),
       addEventListener(_handle, 'transitionstart', (event: TransitionEvent) => {
         if (isAffectingTransition(event)) {
           const animateHandleOffset = () => {